refactor: simplify to single script that lists page IDs
